requirements.txt
Django==1.8.3
Pillow==2.9.0
djangorestframework==3.2.0
django-filter==0.10.0
Markdown==2.6.2